+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] 2024-05-16
+
+### Added
+
+* Added model support for (OpenAI) Gpt4 Omni and replaced the preview versions of GPT4-Turbo / GPT4-Turbo-Vision with
+  GPT-4 Turbo
+* Added model support for (AWS Bedrock) Claude 3 Opus
+
 ## [2.5.4] 2024-05-08
 
 ### Fixed